requirements.txt
anthropic==0.74.1
asgiref==3.8.1
certifi==2024.7.4
channels==4.1.0
daphne==4.1.2
Django==5.2.4
django-bootstrap5==24.3
duckduckgo-search==8.1.1
faiss-cpu==1.9.0.post1
fastapi==0.115.6
flask==3.1.2
grpcio==1.76.0
grpcio-tools==1.76.0
httpx==0.28.0
javalang==0.13.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.79
langchain-ollama==0.2.1
langchain-text-splitters==0.3.11
langgraph==0.2.74
langsmith==0.4.37
markdown==3.7
mcp==1.25.0
nltk==3.9.1
numpy<2.3.0
ollama==0.5.3
openai-whisper==20250625
torch
pillow==12.0.0
protobuf>=6.31.1,<7.0.0
psutil
pydantic==2.11.0
pyinstaller==6.18.0
pymongo==4.16.0
pymupdf==1.26.5
pyodbc==5.3.0
python-docx==1.2.0
python-dotenv==1.0.1
python-pptx==1.0.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-django==4.9.0
pywin32==311
pyyaml==6.0.3
rank-bm25==0.2.2
requests==2.32.3
ruff==0.14.5
setuptools==75.3.0
telethon==1.42.0
typing-extensions==4.12.2
unstructured==0.18.5
uvicorn==0.38.0
websockets==15.0.1
wheel==0.45.0
whitenoise==6.8.2
xhtml2pdf==0.2.17
